test: split up tests into files
